Reconfigure with pseudo-language codes and i18n
Note that psuedo-language codes have been truncated to 8 characters.

i18n (translation) currently breaks with language codes of 9 or more
characters. But this may be easily fixable, and I think is a bug rather
than an enhancement.

However, even with this change, localization (dates, currency, etc.) is
not possible because the psuedo-language codes do not match locales
defined in https://github.com/gohugoio/locales.